Reloading a level now resets the powers and associated gameobjects
Added a reload power function mode
Added list of wasps to Daniel power
A wasp will remove itself from the list of wasps when it hits a peg
Reloading the level while the power is Daniel's will destroy all wasps in the wasp list
Reloading the level while the power is Ethen's will destroy all lines and ensure the player is not in drawing mode
Reloading the level while the power is Kevin's will return the camera to the default position and hide the scope overlay
Reloading the level while the power is Loki's will remove the hook, cord and connection point and ensure that the ball is not considered connected to a peg
Reloading the level while the power is Mateja's will destroy Mateja
Reloading the level while the power is Sweets' will ensure the bucket, launcher and gravity are flipped to the default positions

Programming was completed on the 29/04/2024 (Night of the 28th). Scripts were altered without access to Unity and need to be tested.
Time spent Week of 22-28/04: 00:48:54