Started Creating JackPower
JackPower sets the score gained for each type of peg to be the average score of each active peg and returns the score gained to default values when the power is lost
Added GetAverageActivePegScore Function to PegManager.cs for use in the JackPower

Programming was completed on the 13/04/2024 (Night of the 12th). Scripts were altered without access to Unity and need to be tested.
Time Spent This Week: 00:42:26