Started Creating BenPower
BenPower will spawn an Isaac Prefab instead of a ball on subsequent turns when triggered
Created a RemoveBall function for PlayerControls.cs that handles the destruction of the ball upon it falling out of the play area (or destroying Isaac when he times out)
Restructured the Shoot function to support spawning Isaac instead of the ball when required
Created Isaac.cs that manages the Player's ability to control Isaac's movement, shooting and bomb placement, as well as managing its limited duration

Programming was completed on the 20/04/2024 (Night of the 19th). Scripts were altered without access to Unity and need to be tested.
Time Spent This Week: 00:54:37