Created Dialogue.cs to procedurally fill text boxes with messages
Created SniperBallIndicator.cs to help the player locate the ball when zoomed in with the Kevin Power Sniper Scope
Made the ball gameobject in PlayerControls accessible from other scripts

Programming was completed on the 10/06/2024 (Night of the 9th). Scripts were altered without access to Unity and need to be tested.
Time Spent This Week: 00:41:57