Started Work On Phoebe Power
Reordered Variables in PlayerControls.cs
Fixed bug where power would resolve if it ran out of charges but regained them before the resolve function was triggered

Programming was completed on the 24/06/2024 (Night of the 23rd). Scripts were altered without access to Unity and need to be tested.
Time Spent This Week: 00:38:21