Changed Bocconcinis to be children of their peg
Reworked Bocconcini.cs to have each bocconcini be a child of the peg it replaces so it conforms to the same movement

Renamed SniperBallIndicator.cs to RotateToBall.cs and modified it so it is more generically useable, such as for the eyes of the player characters that track the ball, or the cursor if the ball isn't present

Programming was completed on the 08/07/2024 (Night of the 7th). Scripts were altered without access to Unity and need to be tested.
Time Spent This Week: 00:33:18