Created SaveFile.cs
Started work on SaveFile.cs to manage a user's save file, storing highscores from each level.
SaveFile.cs can read a text file and write to it, storing the high score save data in a public array

Programming was completed on the 22/07/2024 (Night of the 21st). Scripts were altered without access to Unity and need to be tested.
Time Spent This Week: 00:48:40