SaveFile.cs now stores highscores of the same level set on one line
RotateToBall.cs now stores a public bool to indicate whether the object is targeting the ball or the cursor
CenterEyes.cs now only centers the eyes if the eyes were tracking the cursor in addition to the previous requirements
Added Scroll Wheel support to LauncherRotation.cs for minor adjustments to launcher rotation

Programming was completed on the 05/08/2024 (Night of the 4th). Scripts were altered without access to Unity and need to be tested.
Time Spent This Week: 00:57:27