Expanded Dialogue.cs
Added Support for a dialogue screen to UIManager.cs
Dialogue.cs now stores each dialogue sequence
Messages in Dialogue.cs now have an associated character that is saying the message, as well as an expression that appears along side the message
Added option to have all messages generate at the same speed, regardless of length
Added option to automatically load the next level when a dialogue is complete

Programming was completed on the 19/08/2024 (Sleep Schedule was in the process of adjusting at this time, as such programming was completed Monday evening but was still part of the night of the Sunday prior). Scripts were altered without access to Unity and need to be tested.
Time Spent This Week: 00:44:23